\name{PNAR-internal}
\alias{logl_lin}
\alias{scor_lin}
\alias{scor_hessian_out}
\alias{logl_lin2}
\alias{scor_lin2}
\alias{scor_hessian_out2}
\alias{log_logl_lin}
\alias{scor_log}
\alias{scor_hessian_out_log}
\alias{log_logl_lin2}
\alias{scor_log2}
\alias{scor_hessian_out_log2}
\alias{scor_hess_outer_stnar1_h0}
\alias{LM_gamma_stnar }
\alias{optim_gamma_stnar}
\alias{scor_hess_outer_tnar1_h0}
\alias{LM_gamma_tnar}
\alias{optim_gamma_tnar}
\alias{logl_linpq}
\alias{scor_linpq}
\alias{scor_hess_outer_linpq}
\alias{logl_log_linpq}
\alias{scor_logpq}
\alias{scor_hess_outer_logpq}
\alias{LM_gama_stnarpq_j}
\alias{LM_gama_tpnarpq_j}
\alias{LM_gama_stnarpq}
\alias{LM_gama_tnarpq}

\title{Internal PNAR functions}
\description{
Internal functions of package \pkg{PNAR}
}
\section{Functions}{
\itemize{
\item logl_lin( ... )
\item scor_lin( ... )
\item scor_hessian_out( ... )
\item logl_lin2( ... )
\item scor_lin2( ... )
\item scor_hessian_out2( ... )
\item log_logl_lin( ... )
\item scor_log( ... )
\item scor_hessian_out_log( ... )
\item log_logl_lin2( ... )
\item scor_log2( ... )
\item scor_hessian_out_log2( ... )
\item scor_hess_outer_stnar1_h0( ... )
\item LM_gamma_stpnar( ... )
\item optim_gamma_stnar( ... )
\item scor_hess_outer_tnar1_h0( ... )
\item LM_gamma_tnar( ... )
\item optim_gamma_tnar( ... )
\item logl_linpq( ... )
\item scor_linpq( ... )
\item scor_hess_outer_linpq( ... )
\item logl_log_linpq( ... )
\item scor_logpq( ... )
\item scor_hess_outer_logpq( ... )
\item LM_gama_stnarpq_j( ... )
\item LM_gama_tpnarpq_j( ... )
\item LM_gama_stnarpq( ... )
\item LM_gama_tnarpq( ... )
}
}
\details{
These functions are only for internal usage of the \pkg{PNAR} package and not to be called by the user.
}

