\name{PNAR-package}
\alias{PNAR-package}
\alias{PNAR}
\docType{package}
\title{
Poisson Network Autoregressive Models
}
\description{
Computing and plotting the distance covariance and correlation
function of a univariate or a multivariate time series.
}
\details{
\tabular{ll}{
Package: \tab PNAR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2022-04-14\cr
License: \tab GPL(>=2)\cr
}
}
\author{
Michail Tsagris, Mirko Armillotta and Konstantinos Fokianos.
}
\references{
Edelmann, D, K. Fokianos. and M. Pitsillou. (2019). An Updated Literature
Review of Distance Correlation and Its Applications to Time Series.
\emph{International Statistical Review}, 87, 237-262.

Fokianos, K. and M. Pitsillou. (2018). Testing independence for multivariate
time series via the auto-distance correlation matrix.
\emph{Biometrika}, 105, 337-352.

Fokianos, K. and M. Pitsillou. (2017). Consistent testing for pairwise
dependence in time series.
\emph{Technometrics}, 159, 262-3270.
}

\note{
\bold{Disclaimer:} Dr Mirko Armillotta and Konstantinos Fokianos wrote the initial functions.
Dr Tsagris modified them, created the package and is the maintainer of it.
}

\keyword{ package }


